% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{citation.contoso}
\alias{citation.contoso}
\title{Citation for Contoso Package}
\usage{
citation.contoso()
}
\value{
A citation string for the Contoso package
}
\description{
Citation for Contoso Package
}
