% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{create_data_raw}
\alias{create_data_raw}
\title{Create data-raw files}
\usage{
create_data_raw()
}
\value{
list of csv files
}
\description{
Creates the data-raw files of the 10K row size Contoso files if they
they do not already exists.

If it exists it will return nothing
}
