% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_contoso_duckdb}
\alias{create_contoso_duckdb}
\title{Creates duckdb versions of Contoso datasets}
\usage{
create_contoso_duckdb(dir = "temp")
}
\arguments{
\item{dir}{A string indicating the type of DuckDB database to create. The options are:
- `"temp"`: Creates a temporary database stored on disk in a temporary directory.
- `"in_memory"`: Creates an in-memory database that will not persist after the R session ends.
Default is `"temp"`.}
}
\value{
A list of `tbl` objects that are references to the Contoso datasets stored in the DuckDB database. The list contains the following tables:
- `sales`
- `product`
- `customer`
- `store`
- `fx`
}
\description{
Creates duckdb versions of Contoso datasets
}
\details{
The [create_contonso_duckdb] function registers the following Contoso datasets as DuckDB tables:
- `sales`: Contains sales transaction data.
- `product`: Contains details about products, including attributes like product name, manufacturer, and category.
- `customer`: Contains customer demographic and geographic information.
- `store`: Contains information about store locations and attributes.
- `fx`: Contains foreign exchange rate data for currency conversion.
- `date`: Contains various date-related information, including day, week, month, and year.

You can choose to store the database in memory or in a temporary directory. If you choose "temp", the database will be created in a temporary file on disk. If you choose "in_memory", the database will be created entirely in memory and will be discarded after the R session ends.
}
\examples{
# Create a DuckDB version of Contoso datasets stored in a temporary directory
create_contoso_duckdb(dir = "temp")

# Create a DuckDB version of Contoso datasets stored in memory
create_contoso_duckdb(dir = "in_memory")
}
