% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_contoso_duckdb}
\alias{create_contoso_duckdb}
\title{Creates duckdb versions of Contoso datasets}
\usage{
create_contoso_duckdb(db_dir = c("in_memory"), size = "100K")
}
\arguments{
\item{db_dir}{"temp" or "in_memory"}

\item{size}{"100k","1M" or "10M"}
}
\value{
A list of lazy `tbl` objects that are references to the Contoso datasets stored in the DuckDB database. The list contains the following tables:
- `sales`
- `product`
- `customer`
- `store`
- `fx`
- `store`
- `orderrows`
- `date`
}
\description{
Creates duckdb versions of Contoso datasets
}
\details{
The [create_contonso_duckdb] function registers the following Contoso datasets as DuckDB tables:
- `sales`: Contains sales transaction data.
- `product`: Contains details about products, including attributes like product name, manufacturer, and category.
- `customer`: Contains customer demographic and geographic information.
- `store`: Contains information about store locations and attributes.
- `fx`: Contains foreign exchange rate data for currency conversion.
- `date`: Contains various date-related information, including day, week, month, and year.

You can choose to store the database in memory or in a temporary directory. If you choose "temp", the database will be created in a temporary file on disk. If you choose "in_memory", the database will be created entirely in memory and will be discarded after the R session ends.
}
\examples{
# Create a DuckDB version of Contoso datasets stored in memory
create_contoso_duckdb(db_dir = "in_memory",size="100K")
}
