% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fx}
\alias{fx}
\title{Foreign Exchange Data from the Contoso Dataset}
\format{
A data frame with 4 columns:
\describe{
\item{date}{\code{Date} The date of the exchange rate.}
\item{from_currency}{\code{character} The code of the source currency (e.g., "USD", "EUR").}
\item{to_currency}{\code{character} The code of the target currency (e.g., "GBP", "JPY").}
\item{exchange}{\code{double} The exchange rate between the source and target currencies on the given date.}
}
}
\source{
https://github.com/sql-bi/Contoso-Data-Generator-V2-Data/releases/tag/ready-to-use-data
}
\usage{
fx
}
\description{
This dataset contains information about foreign exchange (FX) rates between different currencies. It includes details about the
exchange rate for a given date, as well as the currencies involved. This dataset is useful for analyzing currency conversions
and understanding the exchange rates between different currencies over time.
}
\keyword{datasets}
