% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orders}
\alias{orders}
\title{Order Data from the Contonso Dataset}
\format{
A data frame with 5 columns:
\describe{
\item{order_key}{\code{double} Unique identifier for the order.}
\item{customer_key}{\code{double} Unique identifier for the customer who placed the order.}
\item{store_key}{\code{double} Unique identifier for the store where the order was placed.}
\item{order_date}{\code{Date} The date when the order was placed.}
\item{delivery_date}{\code{Date} The date when the order is expected to be delivered.}
\item{currency_code}{\code{character} The currency code used for the order (e.g., USD, EUR).}
}
}
\source{
https://github.com/sql-bi/Contoso-Data-Generator-V2-Data/releases/tag/ready-to-use-data
}
\usage{
orders
}
\description{
This dataset contains information about customer orders, including order dates, delivery dates, and store details.
}
\keyword{datasets}
